+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Move codebase to github
 - Move doc to readthedocs.io #6
+- Upgrade fiftyone from 0.23 to 0.25 #7
+
+### Fixed
+
+- Fix the fiftyone web app crashing when filtering on attributes columns #7
 
 ## Archived changelog